修复若干buf。。。

+++ src/main/java/com/wip/service/meta/MetaService.java
@@ -79,7 +79,7 @@ public interface MetaService {
      * @param type
      * @return
      */
-    Long getMetasCountByType(String type);
+    int getMetasCountByType(String type);
 
     /**
      * 通过Meta名查找Meta项目
